requirements.txt
# Python 3.x requirements for GPT server
flask>=2.0.0
groq>=0.4.0
python-dotenv>=0.19.0
numpy>=1.19.0
opencv-python>=4.5.0
SpeechRecognition>=3.8.0
requests>=2.25.0
